+++ src/app/src/main/kotlin/org/imperial/mrc/hint/controllers/LoginController.kt
@@ -8,25 +8,18 @@ import org.springframework.web.bind.annotation.GetMapping
 import javax.servlet.http.HttpServletRequest
 import org.imperial.mrc.hint.AppProperties
 import org.imperial.mrc.hint.security.oauth2.OAuth2AuthenticationRedirection
-import org.imperial.mrc.hint.security.oauth2.OAuth2State
 import org.springframework.http.ResponseEntity
 
 @Controller
 class LoginController(
     private val request: HttpServletRequest,
-    private val session: Session,
+    session: Session,
     appProperties: AppProperties,
-    oauth2State: OAuth2State
-) : OAuth2AuthenticationRedirection(appProperties, oauth2State)
+) : OAuth2AuthenticationRedirection(appProperties, session)
 {
     @GetMapping("/login")
-    fun login(model: Model): Any // Return type is string for formLogin, or ResponseEntity for OAuth2
+    fun login(model: Model): String
     {
-        if (appProperties.oauth2LoginMethod)
-        {
-            return oauth2LoginRedirect()
-        }
-
         model["oauth2LoginMethod"] = appProperties.oauth2LoginMethod
         model["title"] = "Login"
         model["username"] = request.getParameter("username") ?: ""
@@ -53,14 +46,20 @@ class LoginController(
             appProperties.applicationTitle
         }
         model["continueTo"] = redirectTo ?: "/"
-        session.setRequestedUrl(redirectTo)
+        session?.setRequestedUrl(redirectTo)
 
         return "login"
     }
 
     @GetMapping("/oauth2")
-    fun loginRedirection(model: Model): ResponseEntity<String>
+    fun loginRedirection(): ResponseEntity<String>
     {
         return oauth2LoginRedirect()
+    }
+
+    @GetMapping("/register")
+    fun registerRedirection(): ResponseEntity<String>
+    {
+        return oauth2RegisterRedirect()
     }
 }