+++ src/app/src/main/kotlin/org/imperial/mrc/hint/security/oauth2/OAuth2LogoutActionBuilder.kt
@@ -1,6 +1,5 @@
 package org.imperial.mrc.hint.security.oauth2
 
-import org.imperial.mrc.hint.AppProperties
 import org.pac4j.core.context.WebContext
 import org.pac4j.core.context.session.SessionStore
 import org.pac4j.core.exception.http.RedirectionAction
@@ -10,9 +9,8 @@ import org.pac4j.core.util.HttpActionHelper
 import java.util.*
 
 class OAuth2LogoutActionBuilder(
-    appProperties: AppProperties,
-    oauth2State: OAuth2State,
-) : LogoutActionBuilder, OAuth2AuthenticationRedirection(appProperties, oauth2State)
+    private val authRedirection: OAuth2AuthenticationRedirection
+) : LogoutActionBuilder
 {
     override fun getLogoutAction(
         context: WebContext?,
@@ -21,6 +19,6 @@ class OAuth2LogoutActionBuilder(
         targetUrl: String?,
     ): Optional<RedirectionAction>
     {
-        return Optional.of(HttpActionHelper.buildRedirectUrlAction(context, oauth2LogoutRedirect()))
+        return Optional.of(HttpActionHelper.buildRedirectUrlAction(context, authRedirection.oauth2LogoutRedirect()))
     }
 }