+++ src/app/src/main/kotlin/org/imperial/mrc/hint/security/oauth2/clients/OAuth2Client.kt
@@ -2,9 +2,9 @@ package org.imperial.mrc.hint.security.oauth2.clients
 
 import org.imperial.mrc.hint.AppProperties
 import org.imperial.mrc.hint.ConfiguredAppProperties
+import org.imperial.mrc.hint.security.oauth2.*
 import org.imperial.mrc.hint.db.UserRepository
 import org.imperial.mrc.hint.security.oauth2.OAuth2LogoutActionBuilder
-import org.imperial.mrc.hint.security.oauth2.OAuth2State
 import org.imperial.mrc.hint.security.oauth2.ProfileDefinition
 import org.pac4j.core.http.callback.PathParameterCallbackUrlResolver
 import org.pac4j.oauth.client.OAuth20Client
@@ -14,23 +14,22 @@ import org.pac4j.scribe.builder.api.GenericApi20
 class OAuth2Client(
     private val userRepository: UserRepository,
     private val appProperties: AppProperties = ConfiguredAppProperties(),
-    private val oAuth2Client: OAuth20Client = OAuth20Client(),
-    private val config: OAuth20Configuration = OAuth20Configuration(),
-    private val oauth2State: OAuth2State = OAuth2State()
+    private val oAuth2Client: OAuth20Client = OAuth20Client()
 ) : HintClients
 {
     override fun hintIndirectClient(): OAuth20Client
     {
         return oAuth2Client.apply {
             callbackUrlResolver = PathParameterCallbackUrlResolver()
-            configuration = getConfig()
+            configuration = getConfig(this.configuration)
             name = "oauth2Client"
             callbackUrl = "${appProperties.applicationUrl}/callback/oauth2Client"
-            logoutActionBuilder = OAuth2LogoutActionBuilder(appProperties, oauth2State)
+            logoutActionBuilder = OAuth2LogoutActionBuilder(OAuth2AuthenticationRedirection(appProperties))
+            credentialsExtractor = OAuth2CredentialExtractor(this.configuration, this)
         }
     }
 
-    private fun getConfig(): OAuth20Configuration
+    private fun getConfig(config: OAuth20Configuration): OAuth20Configuration
     {
         return config.apply {
             responseType = OAuth20Configuration.OAUTH_CODE