+++ src/app/src/test/kotlin/org/imperial/mrc/hint/unit/security/oauth2/OAuth2LogoutActionBuilderTests.kt
@@ -2,6 +2,7 @@ package org.imperial.mrc.hint.unit.security.oauth2
 
 import com.nhaarman.mockito_kotlin.mock
 import org.imperial.mrc.hint.ConfiguredAppProperties
+import org.imperial.mrc.hint.security.oauth2.OAuth2AuthenticationRedirection
 import org.imperial.mrc.hint.security.oauth2.OAuth2LogoutActionBuilder
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
@@ -10,20 +11,14 @@ import org.pac4j.jee.context.JEEContext
 
 class OAuth2LogoutActionBuilderTests
 {
-    private val logoutUrl = "https://fakeUrl/v2/logout?client_id=fakeId&returnTo=http://localhost:8080/login"
-
-    @Test
-    fun `can logout and redirect to login page when login with auth0`()
-    {
-        val sut = OAuth2LogoutActionBuilder(ConfiguredAppProperties(), mock())
-
-        assertEquals(sut.oauth2LogoutRedirect(), logoutUrl)
-    }
-
     @Test
     fun `can return redirect action when login with auth0`()
     {
-        val sut = OAuth2LogoutActionBuilder(ConfiguredAppProperties(), mock())
+        val logoutUrl = "https://fakeUrl/v2/logout?client_id=fakeId&returnTo=http://localhost:8080/login"
+
+        val oAuth2Redirection = OAuth2AuthenticationRedirection(ConfiguredAppProperties(), null)
+
+        val sut = OAuth2LogoutActionBuilder(oAuth2Redirection)
 
         val context = mock<JEEContext>()