+++ src/app/static/src/tests/load/mutations.test.ts
@@ -87,4 +87,10 @@ describe("Load mutations", () => {
         expect(testState.statusPollId).toEqual(-1);
     });
 
+    it("StartPreparingRehydrate sets preparing state", () => {
+        const testState = mockLoadState();
+        mutations.StartPreparingRehydrate(testState, true);
+        expect(testState.loadingState).toBe(LoadingState.SettingFiles);
+    });
+
 });